… vendor/golang.org/x/tools/container/intsets/popcnt_amd64.go
… vendor/golang.org/x/tools/container/intsets/popcnt_amd64.s
… vendor/golang.org/x/tools/container/intsets/popcnt_gccgo.go
… vendor/golang.org/x/tools/container/intsets/popcnt_gccgo_c.c
… vendor/golang.org/x/tools/container/intsets/popcnt_generic.go
… vendor/golang.org/x/tools/container/intsets/sparse.go
… vendor/golang.org/x/tools/container/intsets/util.go