### Issues associated with this PR

https://github.com/mosn/mosn/issues/2008

### Solutions

https://github.com/mosn/mosn/issues/2008#issuecomment-1097844372

### Benchmark

```

func BenchmarkSubsetLoadBalancer(b *testing.B) {
	ps := createHostset(benchHostConfigs(8000, 3))
	subsetConfig := benchSubsetConfig()
	b.Run("subsetLoadBalancer", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			newSubsetLoadBalancer(types.RoundRobin, ps, newClusterStats("BenchmarkSubsetLoadBalancer"), NewLBSubsetInfo(subsetConfig))
		}
	})
}
```

#### before
```
 dzdx@B-QBDRMD6M-0201  ~/Documents/mosn-open/pkg/upstream/cluster   master ±  go test -bench=^BenchmarkSubsetLoadBalancer . -run=^$ -benchmem --cpuprofile=mosn.profile
2022-04-13 18:03:09,468 [INFO] [network] [ register pool factory] register protocol: mock factory
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/upstream/cluster
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkSubsetLoadBalancer/subsetLoadBalancer-12                     12          95145790 ns/op        10267508 B/op       7266 allocs/op
PASS
ok      mosn.io/mosn/pkg/upstream/cluster       2.498s
```
#### after
```
 dzdx@B-QBDRMD6M-0201  ~/Documents/mosn-open/pkg/upstream/cluster   perf/subsetlb ±  go test -bench=^BenchmarkSubsetLoadBalancer . -run=^$ -benchmem --cpuprofile=mosn.profile
2022-04-13 18:03:57,161 [INFO] [network] [ register pool factory] register protocol: mock factory
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/upstream/cluster
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkSubsetLoadBalancer/subsetLoadBalancer-12                    249           5034284 ns/op         2407916 B/op      11464 allocs/op
PASS
ok      mosn.io/mosn/pkg/upstream/cluster       2.626s
```
<img width="1579" alt="image" src="https://user-images.githubusercontent.com/7933207/163155137-c6b72ed7-6cf3-4043-b257-ec989d8f216b.png">

|   | cpu time |  mem alloc  |
|---|---|---|
|  before | 95ms  | 10MB   | 
|  after |  5ms  | 2MB  |


### Code Style
+ Make sure `Goimports` has run
+ Show `Golint` result